+++ CHANGELOG.md
@@ -11,6 +11,8 @@ Bug fixes:
 
 - Resolved an issue with lazy loading introduced in v46.
 - Improve the tab bar rendering with a custom painter (faster and contents are positioned more accurately).
+- Dragging a file from the mindspace tree that is already open in a column will now correctly open the tab in the correct
+  location.
 
 Internal structure changes: